restructured executor slurm
