+++ scripts/performance/llm/finetune_llama31_405b.py
@@ -32,9 +32,14 @@ from ..helpers import (
     get_user_configs,
     set_exp_logging_configs,
     set_primary_perf_configs,
-    slurm_executor,
 )
-from ..utils import get_comm_overlap_callback_idx, hf_tokenizer, import_ckpt_experiment, isfile_train_pack_metadata
+from ..executors import slurm_executor
+from ..utils import (
+    get_comm_overlap_callback_idx,
+    hf_tokenizer,
+    import_ckpt_experiment,
+    isfile_train_pack_metadata,
+)
 
 HF_MODEL_URI = "meta-llama/Llama-3.1-405B"
