+++ scripts/performance/llm/finetune_llama3_8b.py
@@ -25,12 +25,14 @@ from ..argument_parser import parse_cli_args
 from ..helpers import (
     args_sanity_check,
     get_user_configs,
+    set_exp_logging_configs,
+    set_primary_perf_configs,
+)
+from ..executors import slurm_executor
+from ..utils import (
     hf_tokenizer,
     import_ckpt_experiment,
     isfile_train_pack_metadata,
-    set_exp_logging_configs,
-    set_primary_perf_configs,
-    slurm_executor,
 )
 
 HF_MODEL_URI = "meta-llama/Meta-Llama-3-8B"
