--- scripts/performance/recommended_model_configs/strong_scaling_model_configs_h100.csv
@@ -1,11 +1,0 @@
-task,model,size,system,dtype,num_gpus,seq_len,tp_size,pp_size,cp_size,ep_size,vp_size,mbs,gbs,etp_size,cuda_graphs,use_mcore_fsdp,recompute_layers,activation_offload_layers
-pre_train,nemotron4,15b,h100,fp8,128.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,15b,h100,fp8,256.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,15b,h100,fp8,64.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,15b,h100,fp8,16.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,15b,h100,fp8,512.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,15b,h100,fp8,1024.0,4096.0,4.0,1.0,1.0,1.0,1.0,4.0,1024.0,,0,0,0,0
-pre_train,nemotron4,340b,h100,fp8,128.0,4096.0,8.0,8.0,1.0,1.0,12.0,1.0,512.0,,0,0,0,0
-pre_train,nemotron4,340b,h100,fp8,256.0,4096.0,8.0,8.0,1.0,1.0,12.0,1.0,512.0,,0,0,0,0
-pre_train,nemotron4,340b,h100,fp8,512.0,4096.0,8.0,8.0,1.0,1.0,12.0,1.0,512.0,,0,0,0,0
-pre_train,nemotron4,340b,h100,fp8,1024.0,4096.0,8.0,8.0,1.0,1.0,12.0,1.0,512.0,,0,0,0,0
