+++ scripts/performance/vlm/finetune_neva_8b.py
@@ -26,8 +26,8 @@ from ..helpers import (
     get_user_configs,
     set_exp_logging_configs,
     set_primary_perf_configs,
-    slurm_executor,
 )
+from ..executors import slurm_executor
 
 
 def override_recipe_configs(
